HTML: Clean instances of multiple spaces, properly handle multiple bools
HTML_CleanDocument before was not even sending the updated HTML content
data back to the host, so that was addressed on top of checking
occurences of multiple spaces (a ' ' immediately after another ' ').

HTML_ParseAttribute was updated to properly detect a space without a '='
following it is also a boolean attribute, and needs sent to
HTML_ParseAttributeContent accordingly.