Utility: Improve `Util_CharDeleteAtIndex` performance
Before, this utility function would allocate a new string and iterate
through it entirely to fill every entry aside from the index we wanted
to ignore. Now, it simply jumps to the index at hand, and just replaces
it and subsequence indexes with the character in front of it. This
mitigates the need to continually allocate and free memory, instead
relying on the same string build and modifying it. Fixes #6.